Function shipcat::cluster::mass_vault

pub async fn mass_vault<'_, '_>(
    conf: &'_ Config, 
    reg: &'_ Region, 
    n_workers: usize
) -> Result<()>

Apply all vault policies in a region

Generates and writes policies direct to vault using their github team name as auth mappers. Equivalent to:

for team in config.teams:
  shipcat get vaultpolicy {team.name} | vault policy write {team.admins} -
  vault write auth/github/map/teams/{team.admins} value={team.admins}

using vault setup for the vault specified in the Region. If one vault is reused for all regions, this can be done once.

Requires a vault login outside of this command as a user who is sufficiently elevated to write general policies.