Struct shipcat::config::Region

pub struct Region {
    pub namespace: String,
    pub environment: String,
    pub versioningScheme: VersionScheme,
    pub base_urls: BTreeMap<String, String>,
    pub env: BTreeMap<String, String>,
    pub kong: KongConfig,
    pub ip_whitelist: Vec<String>,
    pub kafka: KafkaConfig,
    pub vault: VaultConfig,
    pub locations: Vec<String>,
}

A region is an abstract kube context

Either it's a pure kubernetes context with a namespace and a cluster, or it's an abstract concept with many associated real kubernetes contexts.

Fields
namespace: String

Kubernetes namespace

environment: String

Environment (e.g. dev or staging)

versioningScheme: VersionScheme

Versioning scheme

base_urls: BTreeMap<String, String>

Important base urls that can be templated in evars

env: BTreeMap<String, String>

Environment variables to inject

kong: KongConfig

Kong configuration for the region

ip_whitelist: Vec<String>

List of Whitelisted IPs

kafka: KafkaConfig

Kafka configuration for the region

vault: VaultConfig

Vault configuration for the region

locations: Vec<String>

List of locations the region serves

Methods

impl Region

pub fn secrets(&mut self, region: &str) -> Result<()>

pub fn verify_secrets_exist(&mut self, region: &str) -> Result<()>