Module shipcat::diff

Diffing module for values

Functions

infer_version_change
Infer a version change diff and extract old version and new version

is_version_only
Check if a diff contains only version related changes

minify
Minify diff output from kubectl diff

obfuscate_secrets
Obfuscate a set of secrets from an input string

template_vs_git
Fast local git compare of shipcat template

template_vs_kubectl
Diff using template kubectl diff

values_vs_git
Fast local git compare of the crd

values_vs_kubectl
Diff values using kubectl diff

values_vs_region
Fast local compare of shipcat template for two regions