Module shipcat::kong

A small CLI kong config generator interface

Structs

KongOutput
KongOutput matches the format expected by the Kong Configurator script

KongfigOutput
KongOutput for Kongfig

Enums

KongOutputMode

Functions

config_url
Return the config_url for the given region

generate_kong_output

output
Generate Kong config from a filled in global config