Struct shipcat::manifest::manifest::Manifest

pub struct Manifest {
    pub name: String,
    pub region: String,
    pub environment: String,
    pub namespace: String,
    pub disabled: bool,
    pub external: bool,
    pub regions: Vec<String>,
    pub publiclyAccessible: bool,
    pub secrets: BTreeMap<String, String>,
    pub _decoded_secrets: BTreeMap<String, String>,
    pub chart: Option<String>,
    pub image: Option<String>,
    pub imageSize: Option<u32>,
    pub version: Option<String>,
    pub command: Vec<String>,
    pub metadata: Option<Metadata>,
    pub dataHandling: Option<DataHandling>,
    pub language: Option<String>,
    pub resources: Option<Resources<String>>,
    pub replicaCount: Option<u32>,
    pub env: BTreeMap<String, String>,
    pub configs: Option<ConfigMap>,
    pub vault: Option<VaultOpts>,
    pub httpPort: Option<u32>,
    pub ports: Vec<Port>,
    pub externalPort: Option<u32>,
    pub health: Option<HealthCheck>,
    pub dependencies: Vec<Dependency>,
    pub sidecars: Vec<Sidecar>,
    pub workers: Vec<Worker>,
    pub readinessProbe: Option<Probe>,
    pub livenessProbe: Option<Probe>,
    pub rollingUpdate: Option<RollingUpdate>,
    pub autoScaling: Option<AutoScaling>,
    pub tolerations: Vec<Tolerations>,
    pub hostAliases: Vec<HostAlias>,
    pub volumeMounts: Vec<VolumeMount>,
    pub initContainers: Vec<InitContainer>,
    pub volumes: Vec<Volume>,
    pub cronJobs: Vec<CronJob>,
    pub serviceAnnotations: BTreeMap<String, String>,
    pub labels: BTreeMap<String, String>,
    pub kong: Option<Kong>,
    pub kafka: Option<Kafka>,
    pub secretFiles: BTreeMap<String, String>,
    pub sourceRanges: Vec<String>,
    pub rbac: Vec<Rbac>,
}

Main manifest, serializable from shipcat.yml

Fields
name: String

Name of the service

region: String environment: String namespace: String disabled: bool

Wheter to ignore this service

external: bool

Wheter the service is externally managed

regions: Vec<String>

Regions service is deployed to

publiclyAccessible: bool

Wheter the service should be public

secrets: BTreeMap<String, String> _decoded_secrets: BTreeMap<String, String> chart: Option<String>

Chart to use for the service

image: Option<String>

Optional image name

imageSize: Option<u32>

Optional uncompressed image size (for estimating helm timeouts)

version: Option<String>

Optional version/tag of docker image

command: Vec<String>

Optional image command (if not using the default docker command)

metadata: Option<Metadata>

Canonical data sources like repo, docs, team names

dataHandling: Option<DataHandling>

Data sources and handling strategies

language: Option<String>

Language the service is written in

resources: Option<Resources<String>>

Resource limits and requests

replicaCount: Option<u32>

Replication limits

env: BTreeMap<String, String>

Environment variables to inject

configs: Option<ConfigMap>

Config files to inline in a configMap

vault: Option<VaultOpts>

Vault options

httpPort: Option<u32>

Http Port to expose

ports: Vec<Port>

Ports to open

externalPort: Option<u32>

Externally exposed port

health: Option<HealthCheck>

Health check parameters

dependencies: Vec<Dependency>

Service dependencies

sidecars: Vec<Sidecar>

Sidecars

workers: Vec<Worker>

Worker side-deployments

readinessProbe: Option<Probe>

Optional readiness probe (REPLACES health abstraction)

livenessProbe: Option<Probe>

Optional liveness probe

rollingUpdate: Option<RollingUpdate>

Rolling update Deployment parameters

autoScaling: Option<AutoScaling>

Horizontal Pod Auto Scaler parameters

tolerations: Vec<Tolerations>

Toleration parameters

hostAliases: Vec<HostAlias>

host aliases to inject in /etc/hosts

volumeMounts: Vec<VolumeMount>

Volumes mounts

initContainers: Vec<InitContainer>

Init container intructions

volumes: Vec<Volume>

Volumes

cronJobs: Vec<CronJob>

CronJobs

serviceAnnotations: BTreeMap<String, String>

Service annotations (for internal services only)

labels: BTreeMap<String, String>

Extra labels

kong: Option<Kong>

Kong config

kafka: Option<Kafka>

Kafka config

secretFiles: BTreeMap<String, String>

Kube Secret Files to append

sourceRanges: Vec<String>

Load balancer source ranges

rbac: Vec<Rbac>

Role-Based Access Control

Methods

impl Manifest

Static reducers over available manifests

pub fn get_versions(
conf: &Config,
region: &str
) -> Result<BTreeMap<String, Version>>

Find the hardcoded versions of services in a region

Services without a hardcoded version are not returned.

pub fn get_images(
conf: &Config,
region: &str
) -> Result<BTreeMap<String, String>>

Find the hardcoded images of services in a region

Services without a hardcoded image will assume the shipcat.conf specific default

pub fn get_codeowners(conf: &Config, region: &str) -> Result<Vec<String>>

Generate codeowner strings for each service based based on team owners

Cross references config.teams with manifest.metadata.team Each returned string is Github CODEOWNER syntax

impl Manifest

pub fn set_version(self, ver: &Option<String>) -> Self

Override version with an optional one from the CLI

pub fn print(&self) -> Result<()>

Print manifest to stdout

pub fn verify(&self, conf: &Config) -> Result<()>

Verify assumptions about manifest

Assumes the manifest has been populated with implicits

impl Manifest

Manifests backed by a manifests directory traverse the filesystem for discovery

pub fn available() -> Result<Vec<String>>

Walk the services directory and return the available services

pub fn verify_secrets_exist(&self, vc: &VaultConfig) -> Result<()>

pub fn fill(&mut self, conf: &Config, region: &str) -> Result<()>

Fill in env overrides and apply merge rules

pub fn completed(service: &str, conf: &Config, region: &str) -> Result<Manifest>

Complete (filled in env overrides and populate secrets) a manifest

pub fn stubbed(service: &str, conf: &Config, region: &str) -> Result<Manifest>

Mostly completed but stubbed secrets version of the manifest

pub fn mocked(service: &str, conf: &Config, region: &str) -> Result<Manifest>

Completed manifest with mocked values

pub fn basic(
service: &str,
conf: &Config,
region: Option<String>
) -> Result<Manifest>

A super base manifest - from an unknown region

impl Manifest

pub fn post_merge_implicits(
&mut self,
conf: &Config,
region: Option<String>
) -> Result<()>

Add implicit defaults to self after merging in region overrides

Should be used by entries that have complex implicit results that can be partially overridden I.e. kong struct, dataHandling structs which both have implicit values

pub fn pre_merge_implicits(
&mut self,
conf: &Config,
region: Option<String>
) -> Result<()>

Add implicit defaults to self before merging in values

Should be used by entries that have simple implicit results based on the config I.e. optional strings, integers etc.

pub fn merge(&mut self, pth: &PathBuf) -> Result<()>

Merge defaults from partial override file

Copies keys from environment files into the current manifest struct by default. One special cases are merged carefully:

env dict (merged by key)

impl Manifest

Calculations done based on values in manifests

These generally assume that verify has passed on all manifests.

pub fn estimate_wait_time(&self) -> u32

Estimate how long to wait for a kube rolling upgrade

Was used by helm, now used by the internal upgrade wait time.

pub fn compute_resource_totals(&self) -> Result<ResourceTotals>

Compute the total resource usage of a service

This relies on the Mul and Add implementations of Resources<f64>, which allows us to do + and * on a normalised Resources struct.

impl Manifest

pub fn inline_configs(&mut self, conf: &Config, region: &str) -> Result<()>

Inline templates in values

pub fn template_evars(&mut self, conf: &Config, region: &str) -> Result<()>

Template evars - must happen before inline templates!