Module shipcat::manifest::template

A renderer of tera templates (jinja style)

Used for small app configs that are inlined in the completed manifests.

Functions

init
Initialise the tera templating engine with templates for a service

one_off
One off template

render
Render convenience function that also trims whitespace

service_bound_renderer
Create a one of boxed template renderer for a service

Type Definitions

ContextBoundRenderer
A function that can render templates for a service