Struct shipcat::product::product::Product

pub struct Product {
    pub name: String,
    pub description: String,
    pub version: Option<String>,
    pub location: String,
    pub locations: Vec<String>,
    pub owner: Option<Owner>,
    pub jira: Option<String>,
    pub services: Vec<Service>,
}

Product manifest, serializable from product.yml

Fields
name: String

Name of the product

description: String

Description of the product

version: Option<String>

Version of product

Disjoint from service versions

location: String

Location injected by constructor

locations: Vec<String>

Locations product is active in

owner: Option<Owner>

Owner of the service

jira: Option<String>

Jira ticket

services: Vec<Service>

Service dependencies

Methods

impl Product

pub fn available() -> Result<Vec<String>>

Walk the products directory and return the available products

pub fn fill(&mut self, conf: &Config, location: &str) -> Result<()>

Fill in env overrides and apply merge rules

pub fn completed(
service: &str,
conf: &Config,
location: &str
) -> Result<Product>

Read and environment merged product manifest

pub fn verify(&self, conf: &Config) -> Result<()>

Verify assumptions about product

Assumes the product has been populated with implicits

impl Product

pub fn post_merge_implicits(
&mut self,
_conf: &Config,
location: Option<String>
) -> Result<()>

Add implicit defaults to self after merging in location overrides

pub fn pre_merge_implicits(&mut self, _conf: &Config) -> Result<()>

Add implicit defaults to self before merging in values

pub fn merge(&mut self, pth: &PathBuf) -> Result<()>

Merge defaults from partial override file

Copies keys from environment files into the current product struct by default.