Module shipcat::structs::kafkaresources

Structs

AclDefinition

KafkaResources

KafkaTopics

KafkaUsers

Enums

KafkaUserOperation
Operations relating to a Kafka ACL to be applied onto a resource, values derived from the Strimzi Kafka User Custom Resource Definition Strimzi Kafka User CRD

KafkaUserPatternType

KafkaUserResourceType
Resource Types relating to a Kafka ACL to be applied onto a resource, values derived from the Strimzi Kafka User Custom Resource Definition Strimzi Kafka User CRD