Struct shipcat::structs::kong::Kong

pub struct Kong {
    pub name: String,
    pub upstream_url: String,
    pub unauthenticated: bool,
    pub internal: bool,
    pub publiclyAccessible: bool,
    pub cookie_auth: bool,
    pub cookie_auth_csrf: bool,
    pub uris: Option<String>,
    pub hosts: Option<String>,
    pub host: Option<String>,
    pub auth: Authentication,
    pub strip_uri: bool,
    pub preserve_host: bool,
    pub cors: Option<Cors>,
    pub additional_internal_ips: Vec<String>,
    pub babylon_auth_header: Option<BabylonAuthHeader>,
    pub oauth2_anonymous: Option<String>,
    pub oauth2_extension_plugin: Option<bool>,
    pub upstream_connect_timeout: Option<u32>,
    pub upstream_send_timeout: Option<u32>,
    pub upstream_read_timeout: Option<u32>,
    pub add_headers: Option<BTreeMap<String, String>>,
}

Kong setup for a service

Fields
name: String

Auto-populated name of service

Set internally, and used for defaults (discards value in manifest) But overrideable for extra_apis in Region::kong.

upstream_url: String

The base target URL that points to your API server.

This URL will be used for proxying requests. For example: https://example.com.

Normal kubernetes value is: raftcat.svc.cluster.local If left blank, this value will be generated with the service name instead of raftcat.

unauthenticated: bool

Whether the oauth2 plugin will be applied or not to this api

internal: bool

Whether or not to apply the ip whitelisting (?)

publiclyAccessible: bool

Marker for gate to let external traffic in or not

cookie_auth: bool

Whether to allow cookie based authentication for front-end applications

cookie_auth_csrf: bool

Whether or not to CSRF for cookie auths (mattmalones plugin)

uris: Option<String>

Simple path based routing

E.g. /raftcat

hosts: Option<String>

A comma-separated list of domain names that point to your API.

For example: example.com. At least one of hosts, uris, or methods should be specified

This is the full value sent to kong.

host: Option<String>

Convenience way to specify a single host rather than using hosts above

Will create a single hosts output prefixed with the kong base url. Mutually exclusive with hosts.

auth: Authentication

Authentication type

strip_uri: bool

When matching an API via one of the uris prefixes, strip that matching prefix from the upstream URI to be requested.

false => application has to listen on the uris parameter (e.g. /raftcat) true => application has to listen on /, but use prefix agnostic urls everywhere.

preserve_host: bool

Preserves host headers to backend service

When matching an API via one of the hosts domain names, make sure the request Host header is forwarded to the upstream service. Kong's default is false, meaning the upstream Host header will be extracted from the configured upstream_url.

Shipcat assumes a default of true, as the normal use case is to have this enabled.

cors: Option<Cors>

Configuration parameters for Cross Origin Resource Sharing plugin

When set, the plugin is used.

additional_internal_ips: Vec<String>

When internal is set to true, also add allow these ips through

babylon_auth_header: Option<BabylonAuthHeader>

Babyln plugin (Vincent's) for propagating a core-ruby auth header.

Compatibility layer for old-style core-ruby authorization headers. Deprecated.

oauth2_anonymous: Option<String>

Whether or not to allow anonymous users to go through an authenticated API

Has to match the ID of the anonymous user in the kong configuration. See Region::kong::anonymous_consumers

oauth2_extension_plugin: Option<bool>

Whether or not to use the oauth2 extension plugin for this api.

TODO: can probably be deprecated?

upstream_connect_timeout: Option<u32>

The timeout in milliseconds for establishing a connection to your upstream service. Defaults to 6000

upstream_send_timeout: Option<u32>

The timeout in milliseconds between two successive write operations for transmitting a request to your upstream service Defaults to 60000.

upstream_read_timeout: Option<u32>

The timeout in milliseconds between two successive read operations for transmitting a request to your upstream service Defaults to 60000.

add_headers: Option<BTreeMap<String, String>>

Extra headers to append to the response from kong after reverse proxying

I.e. the application will receive these extra headers.

add_headers:
  Cache-Control: no-cache, no-store
  Strict-Transport-Security: max-age=31536000; includeSubDomains; preload;
  X-Content-Type-Options: nosniff
  X-Frame-Options: SAMEORIGIN
  X-XSS-Protection: 1; mode=block

Methods

impl Kong

pub fn implicits(&mut self, svc: String, reg: Region, tophosts: Vec<String>)

pub fn merge(&mut self, other: Kong)

Merge in fields from an override, if they're set

impl Kong

pub fn verify(&self) -> Result<(), Error>