Struct shipcat::structs::Worker

pub struct Worker {
    pub name: String,
    pub command: Vec<String>,
    pub resources: Resources<String>,
    pub replicaCount: u32,
    pub extraEnv: BTreeMap<String, String>,
    pub httpPort: Option<u32>,
    pub readinessProbe: Option<Probe>,
    pub livenessProbe: Option<Probe>,
}

Worker for a service

Essentially a side-car like object that can scale resources separately to the main pods. Useful for services that have one single side service that polls or does some work.

Fields
name: String

Name of the worker

command: Vec<String>

Image command override

resources: Resources<String>

Resource limits and requests

replicaCount: u32

Replication limits

extraEnv: BTreeMap<String, String>

Extra environment variables

httpPort: Option<u32>

Http Port to expose

readinessProbe: Option<Probe>

Optional readiness probe

livenessProbe: Option<Probe>

Optional liveness probe