Struct shipcat::vault::Vault

pub struct Vault { /* fields omitted */ }

Vault client with cached data

Methods

impl Vault

pub fn from_evars() -> Result<Vault>

Initialize using the same evars or token files that the vault CLI uses

pub fn regional(vc: &VaultConfig) -> Result<Vault>

Initialize using VAULT_TOKEN evar + addr in shipcat.conf

pub fn mocked(vc: &VaultConfig) -> Result<Vault>

Initialize using dummy values and return garbage

pub fn masked(vc: &VaultConfig) -> Result<Vault>

Initialize using vault evars, but return masked return values

pub fn list(&self, path: &str) -> Result<Vec<String>>

List secrets

Does a HTTP LIST on the folder a service is in and returns the keys

pub fn read(&self, key: &str) -> Result<String>

Read secret from a Vault via an authenticated HTTP GET (or memory cache)