Crate shipcat

Modules

audit
Audit objects and API caller

cluster
Cluster level operations

config
Master config with cross-region data

gdpr
gdpr lister

get
Various simple reducers

grafana
A REST interface to grafana using reqwest

graph
A graph generator for manifests using petgraph

helm
A small CLI helm interface

kong
A small CLI kong config generator interface

kube
A small CLI kubernetes interface

list
Convenience listers

show
Simple printers

slack
A post interface to slack using slack_hook

statuscake
A small CLI Statuscake config generator interface

structs
Structs for the manifest

validate
Validation methods of manifests post merge

webhooks
Webhook mux/demux

Structs

AuditWebhook
Where / how to send audited events

Config
Main manifest, serializable from shipcat.yml

Error
The Error type.

KongConfig
Kong configuration for a region

Manifest
Main manifest, serializable from shipcat.yml or the shipcat CRD.

Region
A region is an abstract kube context

Team

Enums

ConfigType
Various states a Config can exist in depending on resolution.

ErrorKind
The kind of an error.

VersionScheme
Versioning Scheme used in region

Webhook
Webhook types that shipcat might trigger after actions

Traits

ResultExt
Additional methods for Result, for easy interaction with this crate.

Functions

init
Smart initialiser with safety

Type Definitions

Result
Convenient wrapper around std::Result.