Struct shipcat::manifest::manifest::Manifest

pub struct Manifest {
    pub name: String,
    pub region: String,
    pub environment: String,
    pub namespace: String,
    pub regions: Vec<String>,
    pub publiclyAccessible: bool,
    pub _decoded_secrets: BTreeMap<String, String>,
    pub disabled: bool,
    pub external: bool,
    pub secrets: BTreeMap<String, String>,
    pub chart: Option<String>,
    pub image: Option<String>,
    pub imageSize: Option<u32>,
    pub version: Option<String>,
    pub command: Vec<String>,
    pub metadata: Option<Metadata>,
    pub dataHandling: Option<DataHandling>,
    pub language: Option<String>,
    pub resources: Option<Resources<String>>,
    pub replicaCount: Option<u32>,
    pub env: BTreeMap<String, String>,
    pub secretFiles: BTreeMap<String, String>,
    pub configs: Option<ConfigMap>,
    pub vault: Option<VaultOpts>,
    pub httpPort: Option<u32>,
    pub ports: Vec<Port>,
    pub externalPort: Option<u32>,
    pub health: Option<HealthCheck>,
    pub dependencies: Vec<Dependency>,
    pub workers: Vec<Worker>,
    pub sidecars: Vec<Sidecar>,
    pub readinessProbe: Option<Probe>,
    pub livenessProbe: Option<Probe>,
    pub rollingUpdate: Option<RollingUpdate>,
    pub autoScaling: Option<AutoScaling>,
    pub tolerations: Vec<Tolerations>,
    pub hostAliases: Vec<HostAlias>,
    pub initContainers: Vec<InitContainer>,
    pub volumes: Vec<Volume>,
    pub volumeMounts: Vec<VolumeMount>,
    pub cronJobs: Vec<CronJob>,
    pub serviceAnnotations: BTreeMap<String, String>,
    pub labels: BTreeMap<String, String>,
    pub kong: Option<Kong>,
    pub hosts: Vec<String>,
    pub kafka: Option<Kafka>,
    pub sourceRanges: Vec<String>,
    pub rbac: Vec<Rbac>,
}

Main manifest, serializable from shipcat.yml

Fields
name: String

Name of the service

This must match the folder name in a manifests repository.

region: String

Region injected into helm chart

Exposed from shipcat, but not overrideable.

environment: String

Environment injected into the helm chart

Exposed from shipcat, but not overrideable.

namespace: String

Namespace injected in helm chart

Exposed from shipcat, but not overrideable.

regions: Vec<String>

Regions to deploy this service to.

Every region must be listed in here. Uncommenting a region in here will partially disable this service.

publiclyAccessible: bool

Wheter the service should be public

This is a special flag not exposed to the charts at the moment.

_decoded_secrets: BTreeMap<String, String> disabled: bool

Service is disabled

This disallows usage of this service in all regions.

external: bool

Service is external

This cancels all validation and marks the manifest as a non-kube reference only.

secrets: BTreeMap<String, String>

Raw secrets from environment variables.

The env Map fills in secrets in this via the vault client. Manifest::secrets partitions env into env and secrets. See Manifest::env.

This means that this is an internal property that is actually exposed!

chart: Option<String>

Chart to use for the service

All the properties in Manifest are tailored towards our base chart, so this should be overridden with caution.

image: Option<String>

Image name of the docker image to run

This can be left out if imagePrefix is set in the config, and the image name also matches the service name. Otherwise, this needs to be the full image name.

imageSize: Option<u32>

Optional uncompressed image size

This is used to compute a more accurate wait time for rolling upgrades. See Manifest::estimate_wait_time.

version: Option<String>

Version aka. tag of docker image to run

This does not have to be set in "rolling environments", where upgrades re-use the current running versions. However, for complete control, production environments should put the versions in manifests.

Versions must satisfy VersionScheme,::verify.

command: Vec<String>

Command to use for the docker image

This can be left out to use the default image command.

metadata: Option<Metadata>

Important contacts and other metadata for the service

Particular uses:

notifying correct people on upgrades via slack

providing direct links to code diffs on upgrades in slack

dataHandling: Option<DataHandling>

Data sources and handling strategies

An experimental abstraction around GDPR

language: Option<String>

Language the service is written in

This does not provide any special behaviour at the moment.

resources: Option<Resources<String>>

Kubernetes resource limits and requests

Api straight from https://kubernetes.io/docs/concepts/configuration/manage-compute-resources-container/

replicaCount: Option<u32>

Kubernetes replication count

This is set on the Deployment object in kubernetes. If you have autoScaling parameters set, then these take precedence.

env: BTreeMap<String, String>

Environment variables to inject

These have a few special convenience behaviours: "IN_VAULT" values is replaced with value from vault/secret/folder/service/KEY One off tera templates are calculated with a limited template context

IN_VAULT secrets will all be put in a single kubernetes Secret object. One off templates can be put in a Secret object if marked | as_secret.

secretFiles: BTreeMap<String, String>

Kubernetes Secret Files to inject

These have the same special "IN_VAULT" behavior as Manifest::env: "IN_VAULT" values is replaced with value from vault/secret/folder/service/key

Note the lowercase restriction on keys. All secretFiles are expected to be base64 in vault, and are placed into a kubernetes Secret object.

configs: Option<ConfigMap>

Config files to inline in a kubernetes ConfigMap

These are read and templated by tera before they are passed to helm. A full tera context from Manifest::make_template_context is used.

vault: Option<VaultOpts>

Vault options

Allows overriding service names and regions for secrets. DEPRECATED. Should only be set in rare cases.

httpPort: Option<u32>

Http Port to expose in the kubernetes Service

This is normally the service your application listens on. Kong deals with mapping the port to a nicer one.

ports: Vec<Port>

Ports to open

For services outside Kong, expose these named ports in the kubernetes Service.

externalPort: Option<u32>

Externally exposed port

Useful for LoadBalancer type Service objects.

health: Option<HealthCheck>

Health check parameters

A small abstraction around readinessProbe. DEPRECATED. Should use readinessProbe.

dependencies: Vec<Dependency>

Service dependencies

Used to construct a dependency graph, and in the case of non-circular trees, it can be used to arrange deploys in the correct order.

workers: Vec<Worker>

Worker Deployment objects to additinally include

These are more flexible than sidecars, because they scale independently of the main replicaCount. However, they are considered separate rolling upgrades. There is no guarantee that these switch over at the same time as your main kubernetes Deployment.

sidecars: Vec<Sidecar>

Sidecars to inject into every kubernetes Deployment

Plain sidecars are injected into the main Deployment and all the workers' ones. They scale directly with the sum of replicaCounts.

readinessProbe: Option<Probe>

readinessProbe for kubernetes

This configures the service's health check, which is used to gate rolling upgrades. https://kubernetes.io/docs/tasks/configure-pod-container/configure-liveness-readiness-probes/

This replaces shipcat's Manifest::health abstraction.

livenessProbe: Option<Probe>

livenessProbe for kubernetes

This configures a readinessProbe check, with the instruction to kill on failure. https://kubernetes.io/docs/tasks/configure-pod-container/configure-liveness-readiness-probes/

rollingUpdate: Option<RollingUpdate>

Rolling update Deployment parameters

These tweak the speed and care kubernetes uses when doing a rolling update. https://kubernetes.io/docs/concepts/workloads/controllers/deployment/#rolling-update-deployment This is attached onto the main Deployment

autoScaling: Option<AutoScaling>

HorizontalPodAutoScaler parameters for kubernetes

Passed all parameters directly onto a spec of a kube HPA. https://kubernetes.io/docs/tasks/run-application/horizontal-pod-autoscale/

tolerations: Vec<Tolerations>

Toleration parameters for kubernetes

Bind a service to a particular type of kube Node. https://kubernetes.io/docs/concepts/configuration/taint-and-toleration/

hostAliases: Vec<HostAlias>

Host aliases to inject in /etc/hosts in every kubernetes Pod

https://kubernetes.io/docs/concepts/services-networking/add-entries-to-pod-etc-hosts-with-host-aliases/

initContainers: Vec<InitContainer>

initContainer list for every kubernetes Pod

Allows database connectivity checks to be done as pre-boot init-step. https://kubernetes.io/docs/concepts/workloads/pods/init-containers/

volumes: Vec<Volume>

Volumes that can be mounted in every kubernetes Pod

Supports our subset of: https://kubernetes.io/docs/concepts/storage/volumes/

volumeMounts: Vec<VolumeMount>

Volumes to mount to every kubernetes Pod

Requires the Manifest::volumes entries. https://kubernetes.io/docs/concepts/storage/volumes/

cronJobs: Vec<CronJob>

Cronjobs images to run as kubernetes CronJob objects

Limited usefulness abstraction, that should be avoided. https://kubernetes.io/docs/concepts/workloads/controllers/cron-jobs/

serviceAnnotations: BTreeMap<String, String>

Annotations to set on Service objects

Useful for LoadBalancer type Service objects. Not useful for kong balanced services.

labels: BTreeMap<String, String>

Labels for every kubernetes object

Injected in all top-level kubernetes object as a prometheus convenience. https://kubernetes.io/docs/concepts/overview/working-with-objects/labels/

kong: Option<Kong>

Kong config

A mostly straight from API configuration struct for Kong Work in progress. structs::kongfig contain the newer abstractions.

hosts: Vec<String>

Hosts to override kong hosts

kafka: Option<Kafka>

Kafka config

A small convencience struct to indicate that the service uses Kafka. The chart will inject a few environment variables and a kafka initContainer if this is set to a Some.

sourceRanges: Vec<String>

Load balancer source ranges

This is useful for charts that expose a Service of LoadBalancer type. IP CIDR ranges, which Kubernetes will use to configure firewall exceptions.

rbac: Vec<Rbac>

Role-Based Access Control

A list of resources to allow the service access to use. This is a subset of kubernetes Role::rules parameters.

Methods

impl Manifest

Static reducers over available manifests

pub fn get_versions(
conf: &Config,
region: &str
) -> Result<BTreeMap<String, Version>>

Find the hardcoded versions of services in a region

Services without a hardcoded version are not returned.

pub fn get_images(
conf: &Config,
region: &str
) -> Result<BTreeMap<String, String>>

Find the hardcoded images of services in a region

Services without a hardcoded image will assume the shipcat.conf specific default

pub fn get_codeowners(conf: &Config, region: &str) -> Result<Vec<String>>

Generate codeowner strings for each service based based on team owners

Cross references config.teams with manifest.metadata.team Each returned string is Github CODEOWNER syntax

impl Manifest

pub fn set_version(self, ver: &Option<String>) -> Self

Override version with an optional one from the CLI

pub fn print(&self) -> Result<()>

Print manifest to stdout

pub fn verify(&self, conf: &Config) -> Result<()>

Verify assumptions about manifest

Assumes the manifest has been populated with implicits

impl Manifest

Manifests backed by a manifests directory traverse the filesystem for discovery

pub fn available() -> Result<Vec<String>>

Walk the services directory and return the available services

pub fn verify_secrets_exist(&self, vc: &VaultConfig) -> Result<()>

pub fn fill(&mut self, conf: &Config, region: &str) -> Result<()>

Fill in env overrides and apply merge rules

pub fn completed(service: &str, conf: &Config, region: &str) -> Result<Manifest>

Complete (filled in env overrides and populate secrets) a manifest

pub fn stubbed(service: &str, conf: &Config, region: &str) -> Result<Manifest>

Mostly completed but stubbed secrets version of the manifest

pub fn mocked(service: &str, conf: &Config, region: &str) -> Result<Manifest>

Completed manifest with mocked values

pub fn basic(
service: &str,
conf: &Config,
region: Option<String>
) -> Result<Manifest>

A super base manifest - from an unknown region

impl Manifest

pub fn post_merge_implicits(
&mut self,
conf: &Config,
region: Option<String>
) -> Result<()>

Add implicit defaults to self after merging in region overrides

Should be used by entries that have complex implicit results that can be partially overridden I.e. kong struct, dataHandling structs which both have implicit values

pub fn pre_merge_implicits(
&mut self,
conf: &Config,
region: Option<String>
) -> Result<()>

Add implicit defaults to self before merging in values

Should be used by entries that have simple implicit results based on the config I.e. optional strings, integers etc.

pub fn merge(&mut self, pth: &PathBuf) -> Result<()>

Merge defaults from partial override file

Copies keys from environment files into the current manifest struct by default. One special cases are merged carefully:

env dict (merged by key)

impl Manifest

Calculations done based on values in manifests

These generally assume that verify has passed on all manifests.

pub fn estimate_wait_time(&self) -> u32

Estimate how long to wait for a kube rolling upgrade

Was used by helm, now used by the internal upgrade wait time.

pub fn compute_resource_totals(&self) -> Result<ResourceTotals>

Compute the total resource usage of a service

This relies on the Mul and Add implementations of Resources<f64>, which allows us to do + and * on a normalised Resources struct.

impl Manifest

pub fn inline_configs(&mut self, conf: &Config, region: &str) -> Result<()>

Inline templates in values

pub fn template_evars(&mut self, conf: &Config, region: &str) -> Result<()>

Template evars - must happen before inline templates!